io module tests beginning
